fix(lint): clear cargo clippy -D warnings (ticket 04 step 0)
CI is about to run `cargo clippy --workspace -- -D warnings`, which
failed on four lints. Two were silenced, two fixed:

  error.rs / component.rs  #[allow(dead_code)] on AppError::BadRequest,
                           SlideElement::is_break and NoteKind::SlideStar
                           — deliberately-unused API surface, kept rather
                           than deleted
  spawning.rs              #[allow(clippy::too_many_arguments)] on
                           spawn_note
  seed_songs.rs            collapsible_if rewritten as a Rust 2024
                           let-chain
  sheets.rs:505            get(&k).is_none() -> !contains_key(&k)
  spawning.rs:402          indexing COUNTDOWN_EDGE_COLORS by loop counter
                           replaced with .iter().enumerate(); the array is
                           declared [Color; COUNTDOWN_EDGES], so this is
                           equivalent. Orphaned the COUNTDOWN_EDGES import,
                           now removed.

Verified: `cargo fmt --check` and `cargo clippy --workspace --all-targets
-- -D warnings` both exit 0, `cargo test --workspace` passes 69 tests
across 11 binaries, and `cargo sqlx prepare --check --workspace --
--all-targets` still exits 0.

+++ engine/src/systems/chart_playback.rs
@@ -73,7 +73,7 @@ impl ChartPlayback {
                         bpm,
                     });
                     let per_comma = absolute_comma_length
-                        .unwrap_or_else(|| DEFAULT_BPM as f64 / bpm as f64 / resolution as f64);
+                        .unwrap_or_else(|| DEFAULT_BPM / bpm as f64 / resolution as f64);
                     current_time += per_comma;
                 }
             }
